refactor: prefix component with Nns
