+++ frontend/src/lib/components/accounts/NnsDestinationAddress.svelte
@@ -2,7 +2,7 @@
   import { emptyAddress } from "$lib/utils/accounts.utils";
   import type { Account } from "$lib/types/account";
   import NnsAddress from "./NnsAddress.svelte";
-  import SelectAccount from "./SelectAccount.svelte";
+  import NnsSelectAccount from "./NnsSelectAccount.svelte";
   import { createEventDispatcher, onMount } from "svelte";
 
   export let filterIdentifier: string | undefined = undefined;
@@ -29,7 +29,7 @@
 <!-- Prevent the component to be presented with a scroll offset when navigating between wizard steps -->
 <!-- note about disableSelection: if user is entering an address with the input field, the address is not empty and therefore no account shall be selected -->
 {#if mounted}
-  <SelectAccount
+  <NnsSelectAccount
     on:nnsSelectAccount={onSelectAccount}
     disableSelection={!emptyAddress(address)}
     displayTitle={true}
