+++ frontend/src/lib/components/accounts/NnsSelectAccount.svelte
