+++ frontend/src/lib/modals/canisters/CreateCanisterModal.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { createEventDispatcher, onMount } from "svelte";
-  import SelectAccount from "$lib/components/accounts/SelectAccount.svelte";
+  import NnsSelectAccount from "$lib/components/accounts/NnsSelectAccount.svelte";
   import ConfirmCyclesCanister from "$lib/components/canisters/ConfirmCyclesCanister.svelte";
   import SelectCyclesCanister from "$lib/components/canisters/SelectCyclesCanister.svelte";
   import { NEW_CANISTER_MIN_T_CYCLES } from "$lib/constants/canisters.constants";
@@ -98,7 +98,7 @@
   >
   <svelte:fragment>
     {#if currentStep?.name === "SelectAccount"}
-      <SelectAccount
+      <NnsSelectAccount
         hideHardwareWalletAccounts
         on:nnsSelectAccount={onSelectAccount}
       />
